Moved User POST interface/type/options to separate types file.
